organization layer process added

+++ src/main/java/engineeromer/com/organization/model/User.java
@@ -17,10 +17,10 @@ import java.util.Set;
 @Table(name = "users")
 public class User {
 
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    private long id;
 
     private String fullName;
